requirements.txt
youtube-transcript-api
requests
# Node.js dependencies (for reference, not installable via pip)
axios
dotenv
portia-sdk-python
portia-sdk-python[all]
